+++ ci/build-docs.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+SERVE=no
+if [ "$1" == "serve" ] ; then
+  SERVE=yes
+fi
+
 for util in mdbook-linkcheck gelatyx ; do
   if ! hash $util 2>/dev/null ; then
     cargo install $util --locked
@@ -8,7 +13,7 @@ done
 
 tracked_markdown=$(mktemp)
 trap "rm ${tracked_markdown}" "EXIT"
-git ls-tree -r HEAD --name-only docs | egrep '\.(markdown|md)$' > $tracked_markdown
+find docs -type f | egrep '\.(markdown|md)$' > $tracked_markdown
 
 gelatyx --language lua --file-list $tracked_markdown --language-config ci/stylua.toml
 gelatyx --language lua --file-list $tracked_markdown --language-config ci/stylua.toml --check || exit 1
@@ -33,7 +38,7 @@ function ghapi() {
 python3 ci/subst-release-info.py || exit 1
 python3 ci/generate-docs.py || exit 1
 
-mdbook-linkcheck --standalone docs
+#mdbook-linkcheck --standalone docs
 
 # Adjust path to pick up pip-installed binaries
 PATH="$HOME/.local/bin;$PATH"
@@ -43,7 +48,7 @@ if ! hash pip3 >/dev/null ; then
   PIP=pip
 fi
 
-$PIP install --quiet mkdocs-material mkdocs-git-revision-date-localized-plugin black mkdocs-exclude mkdocs-include-markdown-plugin
+$PIP install --quiet mkdocs-material mkdocs-git-revision-date-localized-plugin black mkdocs-exclude mkdocs-include-markdown-plugin mkdocs-macros-plugin
 if test -n "${CARDS}" ; then
   $PIP install --quiet pillow cairosvg
 fi
@@ -55,7 +60,8 @@ cp "assets/icon/wezterm-icon.svg" docs/favicon.svg
 mkdir -p docs/fonts
 cp assets/fonts/Symbols-Nerd-Font-Mono.ttf docs/fonts/
 
-mkdocs build
-
-# mdbook-mermaid install docs
-# mdbook build docs
+if [ "$SERVE" == "yes" ] ; then
+  mkdocs "$@"
+else
+  mkdocs build
+fi