+++ docs/changelog.md
@@ -4,6 +4,10 @@ hide:
 toc_depth: 3
 ---
 
+<p style="display:none">
+changelog
+</p>
+
 ## Changes
 
 Releases are named using the date, time and git commit hash.
@@ -17,7 +21,8 @@ usually the best available version.
 
 As features stabilize some brief notes about them will accumulate here.
 
-#### Not Yet
+#### Fixed
+* ssh ProxyCommand didn't parse command lines containing `=` correctly. #3307
 
 ### 20230320-124340-559cb7b0